+++ R/00-NAMESPACE.R
@@ -1,11 +1,14 @@
 #' @rawNamespace import(shiny)
 
+#' @importFrom colourpicker colourInput
 #' @importFrom DataTools tryCatchWithWarningsAndErrors
+#' @importFrom ggplot2 element_blank element_text ggtitle theme xlab xlim ylab ylim
 #' @importFrom grDevices dev.off pdf png svg tiff
 #' @importFrom magrittr %>%
+#' @importFrom openxlsx write.xlsx
 #' @importFrom plotly renderPlotly plotlyOutput save_image
 #' @importFrom rjson toJSON
-#' @importFrom openxlsx write.xlsx
 #' @importFrom utils write.table
+#' @importFrom yaml yaml.load_file
 
 NULL