+++ R/02-plotExport.R
@@ -3,6 +3,7 @@
 #' @param id module id
 #' @param label (character) button label
 #'
+#' @return actionButton
 #' @export
 plotExportButton <- function(id, label = "Export Plot") {
   ns <- NS(id)
@@ -15,25 +16,41 @@ plotExportButton <- function(id, label = "Export Plot") {
 #'
 #' @param id namespace id
 #' @param plotFun (reactive) a reactive function returning a plot for export
+#' @param plotType (character) one of "none", "ggplot". Adds the option to format the
+#'  plot before export
 #' @param filename (character) name of file without file extension
 #' @param plotly (logical) set TRUE if plotFun returns a plotly output
 #' @param plotWidth (reactive) default plot width
 #' @param plotHeight (reactive) default plot height
+#' @param initText (list) optional, named list with title definitions, or output of \code{plotTitlesServer}
+#' @param initRanges (list) optional, named list with range definitions, or output of \code{plotRangesServer}
 #'
 #' @export
 plotExportServer <- function(id,
                              plotFun,
+                             plotType = c("none", "ggplot"),
                              filename = sprintf("%s_plot", gsub("-", "", Sys.Date())),
                              plotly = FALSE,
                              plotWidth = reactive(1280),
-                             plotHeight = reactive(800)) {
+                             plotHeight = reactive(800),
+                             initText = NULL,
+                             initRanges = NULL) {
+  plotType <- match.arg(plotType)
+  formatFun <- switch (plotType,
+                       "none" = noExtraFormat,
+                       "ggplot" = formatWrapperGGplot
+  )
+
   moduleServer(id,
                function(input, output, session) {
-                 observeEvent(input$export, {
+                 observe({
+                   if (inherits(initText, "reactivevalues"))
+                     initText <- reactiveValuesToList(initText)
+
                    plotOutputElement <- if (plotly) {
-                     plotlyOutput(session$ns("plotly"))
+                     plotlyOutput(session$ns("exportPlotly"))
                    } else {
-                     plotOutput(session$ns("plot"), height = "300px")
+                     plotOutput(session$ns("exportPlot"), height = "300px")
                    }
 
                    exportTypeChoices <- if (plotly) {
@@ -46,22 +63,42 @@ plotExportServer <- function(id,
                      title = "Export Graphic",
                      footer = modalButton("OK"),
                      plotOutputElement,
-                     selectInput(
-                       session$ns("exportType"), "Filetype",
-                       choices = exportTypeChoices
+                     fluidRow(
+                       column(4,
+                              h4("File"),
+                              selectInput(
+                                session$ns("exportType"), "Filetype",
+                                choices = exportTypeChoices
+                              ),
+                              if (!plotly) numericInput(session$ns("width"), "Width (px)", value = plotWidth()) else NULL,
+                              if (!plotly) numericInput(session$ns("height"), "Height (px)", value = plotHeight()) else NULL,
+                       ),
+                       if (plotType == "ggplot") {
+                         column(4, plotTitlesUI(session$ns("titlesFormat"),
+                                                type = "ggplot",
+                                                initText = initText))
+                       } else NULL,
+                       if (plotType == "ggplot") {
+                         column(4, plotRangesUI(session$ns("axesRanges"),
+                                                initRanges = initRanges))
+                       }
                      ),
-                     if (!plotly) numericInput(session$ns("width"), "Width (px)", value = plotWidth()) else NULL,
-                     if (!plotly) numericInput(session$ns("height"), "Height (px)", value = plotHeight()) else NULL,
+                     tags$br(),
                      downloadButton(session$ns("exportExecute"), "Export"),
                      easyClose = TRUE
                    ))
+                 }) %>%
+                   bindEvent(input$export)
+
+                 text <- plotTitlesServer("titlesFormat", type = plotType, initText = initText)
+                 ranges <- plotRangesServer("axesRanges", type = plotType, initRanges = initRanges)
+
+                 output$exportPlot <- renderPlot({
+                   plotFun()() %>%
+                     formatFun(text = text, ranges = ranges)
                  })
 
-                 output$plot <- renderPlot({
-                   plotFun()()
-                 })
-
-                 output$plotly <- renderPlotly({
+                 output$exportPlotly <- renderPlotly({
                    plotFun()()
                  })
 
@@ -82,12 +119,19 @@ plotExportServer <- function(id,
                               tiff = tiff(file, width = input$width, height = input$height),
                               svg = svg(file, width = input$width / 72, height = input$height / 72)
                        )
-                       print(plotFun()())
+                       print(
+                         plotFun()() %>%
+                           formatFun(text = text, ranges = ranges)
+                       )
                        dev.off()
                      }
                    }
                  )
                })
+}
+
+noExtraFormat <- function(plot, ...) {
+  plot
 }
 
 # TEST MODULE -------------------------------------------------------------
@@ -104,24 +148,52 @@ plotExportServer <- function(id,
 #       collapsible = TRUE,
 #       id = "test"
 #     ),
+#     plotOutput("plot"),
 #     plotExportButton(id = "expPlot")
 #   )
 # )
 #
 # server <- function(input, output, session) {
-#   plotExportServer("expPlot",
-#                    plotFun = reactive({
-#                      function() {
-#                        data <- data.frame(
-#                          x = c(1, 2, 3, 4, 5),
-#                          y = c(2, 4, 1, 7, 3)
-#                        )
+#   testTitles <- reactiveValues(
+#     plotTitle = list(text = "testHeader", fontType = "italic", color = "#000000",
+#                 size = 30, hide = FALSE),
+#     xAxisTitle = list(text = "test", fontType = "bold",
+#                  color = "#FF00EA", size = 25, hide = FALSE),
+#     yAxisTitle = list(text = "", fontType = "plain", color = "#000000",
+#                  size = 12L, hide = FALSE),
+#     xAxisText = list(fontType = "bold",
+#                       color = "#FF00EA", size = 25, hide = FALSE),
+#     yAxisText = list(fontType = "plain", color = "#000000",
+#                       size = 12L, hide = FALSE)
+#   )
 #
-#                        ggplot2::ggplot(data, ggplot2::aes(x = x, y = y)) +
-#                          ggplot2::geom_point()
-#                      }
-#                      }),
-#                    filename = "plot")
+#   testRanges <- reactiveValues(
+#     xAxis = list(min = 0L, max = 20, fromData = TRUE),
+#     yAxis = list(min = 0L, max = 10L, fromData = FALSE)
+#   )
+#
+#   testPlotFun <- function() {
+#     data <- data.frame(
+#       x = c(1, 2, 3, 4, 5),
+#       y = c(2, 4, 1, 7, 3)
+#     )
+#
+#     ggplot2::ggplot(data, ggplot2::aes(x = x, y = y)) +
+#       ggplot2::geom_point()
+#   }
+#
+#   output$plot <- renderPlot({
+#     testPlotFun() %>%
+#       formatTitlesOfGGplot(text = testTitles) %>%
+#       formatRangesOfGGplot(ranges = testRanges)
+#   })
+#
+#   plotExportServer("expPlot",
+#                    plotFun = reactive({ testPlotFun }),
+#                    plotType = "ggplot",
+#                    filename = "plot",
+#                    initText = testTitles,
+#                    initRanges = testRanges)
 # }
 #
 # shinyApp(ui = ui, server = server)